chore: cleanup function signatures (#215)
This resolves INT-4552.

More specifically, this does three things:
- [x] Adds a CI check for matching signatures in `crates/tracegen`
between Rust and CUDA. It already found one number-of-arguments-mismatch
in some sha function, one we-pass-`u32`-but-cuda-reads-`size_t` in field
expression tracegen (although cuda code instantly converted it back to
`uint32_t` so it didn't trigger), and another error of the same sort for
some other function, where a lot of parameters we passed were `u32` but
needed to be `size_t`. It also caught a lot of reverse mismatches, where
we passed `size_t` and received `u32`, which I think was not a big deal,
but I fixed them anyway. The two scripts I mostly got from chat gpt are
now in `tracegen/scripts`.
- [x] Adds a struct that holds a pointer with some size in bytes, aka a
view to a `DeviceBuffer`. In Rust this is just `struct DeviceBufferView
{ ptr: *const c_void, size: usize }`, in CUDA I decided to make two
separate structures `DeviceRawBufferConstView { uintptr_t ptr; size_t
size; }`, and `template <typename T> struct DeviceBufferConstView` that
also provides easy access (aka `operator []`). The `DeviceBufferView` in
Rust has the same alignment etc as those three CUDA structs, because 1)
it does and 2) the script above doesn't complain. A lot of places where
we just passed a pointer and a length or `num_records` now become
simpler, and we do a lot fewer
`reinterpret_cast<RecordType*>(records)[idx]`. This PR doesn't claim to
change all of these places, but probably most of them.
- [x] We now just pass trace buffers as `*mut F` instead of
`tracegen<T>(d_trace: &DeviceBuffer<T>)` or anything. On the receiving
side this is treated as `Fp* d_trace` anyway.

---------

+++ tracegen-gpu/cuda/src/extensions/native/poseidon2/kernels.cu
@@ -397,15 +397,15 @@ __global__ void cukernel_native_poseidon2_tracegen(
 
 extern "C" int _native_poseidon2_tracegen(
     Fp *d_trace,
-    size_t height,
-    size_t width,
+    uint32_t height,
+    uint32_t width,
     Fp *d_records,
-    size_t num_records,
+    uint32_t num_records,
     uint32_t *d_chunk_start,
     uint32_t num_chunks,
     uint32_t *d_range_checker,
     uint32_t range_checker_num_bins,
-    size_t sbox_regs,
+    uint32_t sbox_regs,
     uint32_t timestamp_max_bits
 ) {
     auto [grid, block] = kernel_launch_params(height - num_records + num_chunks, 256);
